Java Class Buttons
This is an unfinished Java class that will most likely be broken up into
other classes at a later date.  It has quite a lot of errors right now
and unfinished methods. I am still familiarizing myself with the Android
library, which has lot of interfaces that are new to me.